backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
Flask-JWT-Extended==4.6.0
pymongo==4.10.1
bcrypt==4.2.0
python-dotenv==1.0.1
stripe==10.12.0
gunicorn==23.0.0
eventlet==0.37.0
reportlab==4.2.5
openpyxl==3.1.5
pillow==11.0.0
